enhance: Migrate useRemember() to signals
Inertia's useRemember() implementation in official adapters give you the possibility to store
simple values, such as numbers or strings, in history and does not only limit to objects
and arrays.

When implementing this helper myself, I was just introduced to Solid's Stores and was looking
forward abusing them for this helper, but it meant that we would need to sacrifice simple
JS values and limit the functionality to objects and arrays.

By migrating the function to signals, we can also make use of simpler values, although we
sacrifice granular reactivited that stores provide, however, I believe it is a fair price
to pay.

If there is demand to bring back a useRemember() helper with granular reactivity, we may
be able to provide a non-standard helper function to provide such functionality.